requirements.txt
appdirs==1.4.4
asn1crypto==1.5.1
attrs==25.4.0
Babel==2.9.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
cached-property==2.0.1
certifi==2026.1.4
cffi==2.0.0
chardet==3.0.4
cryptography==2.6.1
decorator==4.4.2
defusedxml==0.7.1
docutils==0.16
ebaysdk==2.1.5
freezegun==0.3.15
idna==2.8
isodate==0.7.2
Jinja2==2.11.3
libsass==0.18.0
lxml==4.6.5
MarkupSafe==1.1.0
num2words==0.5.6
numpy==2.2.6
ofxparse==0.21
packaging==26.0
pandas==2.2.2
paramiko==2.12.0
passlib==1.7.3
Pillow==9.0.1
polib==1.1.0
psutil==5.6.7
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydot==1.4.1
PyNaCl==1.6.2
pyOpenSSL==19.0.0
pyparsing==3.3.2
PyPDF2==1.26.0
pyserial==3.4
python-dateutil==2.8.2
python-ldap==3.4.0
python-stdnum==1.13
pytz==2025.2
pyusb==1.0.2
qrcode==6.1
reportlab==3.5.59
requests==2.25.1
requests-toolbelt==1.0.0
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.5
vobject==0.9.6.1
Werkzeug==2.0.2
xlrd==1.2.0
XlsxWriter==1.1.2
xlwt==1.3.0
zeep==3.4.0
